[chore] Move spanmetrics cfg validation to ConfigValidator#Validate.
