+++ connector/spanmetricsconnector/config.go
@@ -15,11 +15,15 @@
 package spanmetricsconnector // import "github.com/open-telemetry/opentelemetry-collector-contrib/connector/spanmetricsconnector"
 
 import (
+	"fmt"
 	"time"
 
+	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/featuregate"
 	"go.opentelemetry.io/collector/pdata/pmetric"
 )
+
+var _ component.ConfigValidator = (*Config)(nil)
 
 const (
 	delta      = "AGGREGATION_TEMPORALITY_DELTA"
@@ -70,6 +74,23 @@ type Config struct {
 	Namespace string `mapstructure:"namespace"`
 }
 
+// Validate checks if the processor configuration is valid
+func (c Config) Validate() error {
+	err := validateDimensions(c.Dimensions, dropSanitizationGate.IsEnabled())
+	if err != nil {
+		return err
+	}
+
+	if c.DimensionsCacheSize <= 0 {
+		return fmt.Errorf(
+			"invalid cache size: %v, the maximum number of the items in the cache should be positive",
+			c.DimensionsCacheSize,
+		)
+	}
+
+	return nil
+}
+
 // GetAggregationTemporality converts the string value given in the config into a AggregationTemporality.
 // Returns cumulative, unless delta is correctly specified.
 func (c Config) GetAggregationTemporality() pmetric.AggregationTemporality {
@@ -77,4 +98,33 @@ func (c Config) GetAggregationTemporality() pmetric.AggregationTemporality {
 		return pmetric.AggregationTemporalityDelta
 	}
 	return pmetric.AggregationTemporalityCumulative
+}
+
+// validateDimensions checks duplicates for reserved dimensions and additional dimensions. Considering
+// the usage of Prometheus related exporters, we also validate the dimensions after sanitization.
+func validateDimensions(dimensions []Dimension, skipSanitizeLabel bool) error {
+	labelNames := make(map[string]struct{})
+	for _, key := range []string{serviceNameKey, spanKindKey, statusCodeKey} {
+		labelNames[key] = struct{}{}
+		labelNames[sanitize(key, skipSanitizeLabel)] = struct{}{}
+	}
+	labelNames[spanNameKey] = struct{}{}
+
+	for _, key := range dimensions {
+		if _, ok := labelNames[key.Name]; ok {
+			return fmt.Errorf("duplicate dimension name %s", key.Name)
+		}
+		labelNames[key.Name] = struct{}{}
+
+		sanitizedName := sanitize(key.Name, skipSanitizeLabel)
+		if sanitizedName == key.Name {
+			continue
+		}
+		if _, ok := labelNames[sanitizedName]; ok {
+			return fmt.Errorf("duplicate dimension name %s after sanitization", sanitizedName)
+		}
+		labelNames[sanitizedName] = struct{}{}
+	}
+
+	return nil
 }
